Fix some minor compiler warnings.
git-svn-id: http://svn.openimageio.org/oiio/trunk@350 31904243-a364-42ac-a3db-9305396518ba

+++ src/jpeg.imageio/jpegexif.cpp
@@ -238,7 +238,7 @@ print_dir_entry (const TIFFDirEntry &dir, const char *datastart)
         break;
     case TIFF_UNDEFINED :
     case TIFF_NOTYPE :
-        for (int i = 0;  i < dir.tdir_count;  ++i)
+        for (size_t i = 0;  i < dir.tdir_count;  ++i)
             std::cerr << (int)((unsigned char *)mydata)[i] << ' ';
     default:
         break;
